+++ README.md
@@ -1,0 +1,56 @@
+**<h1 align="center"> **QuizWiz** </h1>**
+
+
+QuizWiz is designed to be a fun mobile centric based web app that allows users to compete for spots on our leaderboard with their High-Scores from the quizzes they take. It is a responsive app that utilizes JWT for authorization and auhtentication, allowing users to login with their Google, Facebook or Github Credentials. Once successfully logged in you are taken to a profile page that also has an auto-playing carousel with the Top Three leading scores.
+User can then navigate to the home page where you choose the category of your quiz. After the quiz is complete and you submit your score, you can navigate to the leaderboard to see the Top-Ten scores. 
+ 
+<br />
+
+<h2>Technologies used</h2>
+<br />
+  
+<div align='center'>
+  <img  width="70" height="70" src="https://github.com/BrandonTaft/BrandonTaft/blob/main/assets/passportjs.png" />
+ &nbsp;&nbsp;&nbsp;
+  <img  width="70" height="70" src="https://github.com/BrandonTaft/BrandonTaft/blob/main/assets/react-original.svg" />
+ &nbsp;&nbsp;&nbsp;
+ <img  width="70"height="70" src="https://github.com/BrandonTaft/BrandonTaft/blob/main/assets/javascript-original.svg" />
+ &nbsp;&nbsp;&nbsp;
+  <img  width="70" height="70" src="https://github.com/BrandonTaft/BrandonTaft/blob/main/assets/bootstrap-original.svg" />
+ &nbsp;&nbsp;&nbsp;
+  <img  width="70" height="70" src="https://github.com/BrandonTaft/BrandonTaft/blob/main/assets/html5-original.svg"  />
+ &nbsp;&nbsp;&nbsp;
+  <img  width="70" height="70" src="https://github.com/BrandonTaft/BrandonTaft/blob/main/assets/css3-original.svg" />
+ &nbsp;&nbsp;&nbsp;
+  <img  width="70" height="70" src="https://github.com/BrandonTaft/BrandonTaft/blob/main/assets/nodejs-original.svg" />
+  &nbsp;&nbsp;&nbsp;
+  <img  width="70" height="70" src="https://github.com/BrandonTaft/BrandonTaft/blob/main/assets/postgresql-original.svg"    />
+ <br />
+ Passport &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;React &nbsp; &nbsp;&nbsp; &nbsp;Javascript&nbsp;&nbsp;&nbsp;&nbsp;  &nbsp;Bootstrap &nbsp;&nbsp;&nbsp;&nbsp; HTML5 &nbsp;&nbsp; &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;CSS&nbsp;&nbsp;&nbsp;&nbsp; &nbsp;&nbsp;&nbsp;&nbsp;&nbsp; Nodejs &nbsp;  PostgresSQL
+ <br />
+ <br />
+
+ 
+ _____
+  
+
+https://user-images.githubusercontent.com/86688411/140661143-ea00f393-fde1-4329-89bb-278e557a897c.mp4
+ 
+  <br />
+
+ <img width="45%" src="https://github.com/BrandonTaft/Quizwiz/blob/main/quiz-log.png" />
+  &nbsp;&nbsp;&nbsp; &nbsp;&nbsp;&nbsp; &nbsp;&nbsp;&nbsp;
+<img width="45%"  src="https://github.com/BrandonTaft/Quizwiz/blob/main/quiz.png" />
+  &nbsp;&nbsp;&nbsp; &nbsp;&nbsp;&nbsp; &nbsp;&nbsp;&nbsp;
+  <br />
+  <br />
+  <br />
+<img width="45%"  src="https://github.com/BrandonTaft/Quizwiz/blob/main/quizhome.png" />
+  &nbsp;&nbsp;&nbsp; &nbsp;&nbsp;&nbsp; &nbsp;&nbsp;&nbsp;
+<img width="45%"  src="https://github.com/BrandonTaft/Quizwiz/blob/main/quizprofile.png" />
+   &nbsp;&nbsp;&nbsp; &nbsp;&nbsp;&nbsp; &nbsp;&nbsp;&nbsp;
+</div>
+
+ <h2 align="center">Thank You For Stopping By!</h2> 
+<p align='center'>For enquiries or empolyment oppurtunities, please reach out at <a href="mailto:btaftcan@gmail.com">btaftcan@gmail.com</a> or over on <a href="https://www.linkedin.com/in/brandonmtaft">LinkedIn</a>.</p>
+
