--- server/categories.txt
@@ -1,12 +1,0 @@
-General Information - 9
-Books - 10
-Movies - 11
-TV - 14
-Music - 12
-Video Games - 15
-Math - 19
-Computers - 18
-Geography - 22
-Sports - 21
-Celebrities - 26
-Random - 100
