--- server/config/config.json
@@ -1,23 +1,0 @@
-{
-  "development": {
-    "username": "zwrxldqy",
-    "password": "EWGVlAaZysUApj06BYCIt3TWAnwFLLPu",
-    "database": "zwrxldqy",
-    "host": "fanny.db.elephantsql.com",
-    "dialect": "postgres"
-  },
-  "test": {
-    "username": "zwrxldqy",
-    "password": "EWGVlAaZysUApj06BYCIt3TWAnwFLLPu",
-    "database": "zwrxldqy",
-    "host": "fanny.db.elephantsql.com",
-    "dialect": "postgres"
-  },
-  "production": {
-    "username": "zwrxldqy",
-    "password": "EWGVlAaZysUApj06BYCIt3TWAnwFLLPu",
-    "database": "zwrxldqy",
-    "host": "fanny.db.elephantsql.com",
-    "dialect": "postgres"
-  }
-}
