--- server/models/users.js
@@ -1,28 +1,0 @@
-'use strict';
-const {
-  Model
-} = require('sequelize');
-module.exports = (sequelize, DataTypes) => {
-  class Users extends Model {
-    /**
-     * Helper method for defining associations.
-     * This method is not a part of Sequelize lifecycle.
-     * The `models/index` file will call this method automatically.
-     */
-    static associate(models) {
-      // define association here
-    }
-  };
-  Users.init({
-    name: DataTypes.STRING,
-    password: DataTypes.STRING,
-    high_score: DataTypes.INTEGER,
-    avatar: DataTypes.STRING,
-    spare_one: DataTypes.STRING,
-    spare_two: DataTypes.TEXT
-  }, {
-    sequelize,
-    modelName: 'Users',
-  });
-  return Users;
-};
