--- server/package.json
@@ -1,45 +1,0 @@
-{
-  "name": "hello-node",
-  "version": "0.0.1",
-  "description": "QuizWiz",
-  "main": "server.js",
-  "scripts": {
-    
-  },
-  "dependencies": {
-    "cors": "^2.8.5",
-    "express": "^4.17.1",
-    "handlebars": "^4.7.7",
-    "pg": "^8.7.1",
-    "point-of-view": "^4.14.0",
-    "sequelize": "^6.6.5",
-    "sequelize-cli": "^6.2.0",
-    "dotenv": "^10.0.0",
-    "bcrypt": "^5.0.1",
-    "bcryptjs": "^2.4.3",
-    "jsonwebtoken": "^8.5.1",
-    "axios": "^0.22.0",
-    "cookie-parser": "^1.4.5",
-    "passport-facebook": "^3.0.0",
-    "passport-google-oauth20": "^2.0.0",
-    "passport-instagram": "^1.0.0",
-    "passport-github2": "^0.1.12",
-    "passport-local": "^1.0.0",
-    "express-session": "^1.17.2",
-    "connect": "^3.7.0",
-    "flash": "^1.1.0",
-    "passport": "^0.5.0",
-    "passport-jwt": "^4.0.0",
-    "uuid": "^8.3.2"
-  },
-  "engines": {
-    "node": "12.x"
-  },
-  "license": "MIT",
-  "keywords": [
-    "node",
-    "express"
-  ],
-  "devDependencies": {},
-  "author": ""
-}
